Delinquent Duo
prova2 post, get e delete prenotazioni funzionanti LEGGERE PROMEMORIA.txt

+++ src/main/java/com/generation/vehiclesharing/entities/Veicolo.java
@@ -43,7 +43,7 @@ public class Veicolo {
 	
 	
 
-	public long getVeicolo_id() {
+	public int getVeicolo_id() {
 		return veicolo_id;
 	}
